pyproject.toml
[tool.poetry]
name = "blink-api"
version = "0.1.0"
description = ""
authors = ["arunabhreckonsys <arunabh@reckonsys.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
black = "^24.4.2"
logtail-python = "^0.3.0"
structlog = "^24.4.0"
opentelemetry-distro = "0.43b0"
opentelemetry-exporter-otlp = "1.22.0"
gevent = "^24.2.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
